backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
schedule==1.2.0
pandas==2.0.3
scikit-learn==1.3.0
